requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (Local SQLite - FREE)
# No external database needed - uses local SQLite

# AI/ML (OpenRouter Free Tier)
# google-generativeai  # PAID - REMOVED

# Web Scraping (FREE)
beautifulsoup4==4.12.3

# Vector DB for RAG (FREE, Local)
chromadb==0.4.22
sentence-transformers==2.3.1

# Data Processing
pandas>=2.2.0
openpyxl==3.1.2
xlsxwriter==3.1.9

# Image Processing & OCR
Pillow==10.2.0
pytesseract==0.3.10

# Email Processing
email-validator==2.1.0
email-reply-parser==0.5.12

# Google Sheets Integration
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# HTTP & Webhooks
httpx>=0.27.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.10.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
